% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article_assignment.R
\name{allocate_randomly_with_some_double_assignment}
\alias{allocate_randomly_with_some_double_assignment}
\title{Assigns articles to users with some double coding and some another coding rate. Some users have a specified number of articles also assigned to another user (double coding). Remaining assignments are randomly allocated across all users.}
\usage{
allocate_randomly_with_some_double_assignment(double_coding_user_ids,
  other_user_ids, set, n_double_coding, coder_rate = 1.1,
  allocation_type = "coding", allocated_by = "irregular_random_assignment",
  restrict_to_actual = TRUE, make_assignments = TRUE)
}
\arguments{
\item{double_coding_user_ids}{The users who will have a specified number of their articles double coded. to allocate the articles amongst (vector with single or multiple user_ids).}

\item{other_user_ids}{The other users to allocate the articles amongst (vector with single or multiple user_ids).}

\item{set}{The set of documents which are to be allocated (vector with single or multiple of document_ids).}

\item{n_double_coding}{Number of articles (per user) that should be doubled coded (by those users undertaking double coding).}

\item{coder_rate}{The average number of coders to be allocated to a document.}

\item{allocation_type}{The value to write to the allocation_type field in the database document_allocations table (training, testing, coding, checking, ideal).}

\item{allocated_by}{The value to write to the allocated_by field in the database document_allocations table.}

\item{restrict_to_actual}{Should the restriction to actual users and documents be enforced. Should only be set to FALSE for debugging purposes.}

\item{make_assignments}{Actually make changes to the database (TRUE) or only create a proposed set of changes as a dataframe (FALSE)}
}
\description{
\code{allocate_randomly_with_some_double_assignment} assigns articles to some users with a specified number of articles to be double coded by the other users. Articles which are not used in this double coding are allocated amongst all users at the specified coding rate.
By default the articles and user ids are checked against the database and codes which do not correspond to existing documents and/or users are ignored.
}
