% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article_assignment.R
\name{assign_article_to_user}
\alias{assign_article_to_user}
\alias{assign_set}
\alias{assign_testset_to_user}
\alias{assign_trainingset_to_user}
\alias{assign_initalsets_to_users}
\title{Assign article to a user.}
\usage{
assign_article_to_user(con, document_id, user_id, allocation_type, allocated_by,
  allocation_date = as.character(Sys.Date()), status = "NEW",
  coding_complete = 0)

assign_set(con, user_id, set, allocation_type, allocated_by = "assign_set",
  status = "NEW")

assign_testset_to_user(con, user_id, allocation_type = "testing",
  allocated_by = "assign_testset_to_user")

assign_trainingset_to_user(con, user_id, allocation_type = "training",
  allocated_by = "assign_trainingset_to_user")

assign_initalsets_to_users(con, user_ids)
}
\arguments{
\item{con}{The database connection to the election violence database.}

\item{document_id}{Id of the document to be assigned or a vector of document ids.}

\item{user_id}{Id of the user the document is to be assigned to or a vector of user ids.}

\item{allocation_type}{Type of allocation (training, testing, coding, checking, ideal).}

\item{allocation_date}{Date allocation made (usually today).}

\item{status}{Status of document coding (generally 'NEW' for newly assigned documents).}
}
\description{
\code{assign_article_to_user} assigns a specific article or set of articles to a specific user or set of users. Other document assignment functions are convenience wrappers around this funciton with arguments set appropriately.

\code{assign_set} assigns a whole set of articles to a user (articles which are already allocated to a user are not allocated again).

\code{assign_testset_to_user} assigns the test set of articles to a user specified by the user id. Articles which are already allocated to a user are not allocated again.

\code{assign_trainingset_to_user} assigns the training set to user (articles which are already allocated to a user are not allocated again).

\code{assign_initalsets_to_users} assigns the training set and test sets to a user or set of users (articles which are already allocated to a user are not allocated again).
}
\section{Functions}{
\itemize{
\item \code{assign_set}: Assign set of articles to user

\item \code{assign_testset_to_user}: Assigns testset to user or users

\item \code{assign_trainingset_to_user}: Assigns training set to user

\item \code{assign_initalsets_to_users}: Assigns training and testset set to user
}}

