% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{assign_autoclusters_to_user}
\alias{assign_autoclusters_to_user}
\title{Assigns mutiple auto detected clusters to a single user.}
\usage{
assign_autoclusters_to_user(autocluster_ids, user_id)
}
\arguments{
\item{autocluster_ids}{The autocluster_ids to allocate.}

\item{user_id}{The user id to allocate to.}
}
\description{
Assigns mutiple auto detected clusters to a single user.
}
