% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{assign_clusters_from_df}
\alias{assign_clusters_from_df}
\title{Assigns event report to an autodetected cluster}
\usage{
assign_clusters_from_df(
  df,
  cluster_col = geodatecluster,
  event_report_id_col = event_report_id
)
}
\arguments{
\item{df}{a data frame or tibble indicating clustering to upload to the database.}

\item{cluster_col}{the column of the data frame which indicates the clustering.}

\item{event_report_id_col}{the column of the data frame which identifies the event report}
}
\description{
Assigns event report to an autodetected cluster
}
