% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{assign_eventreport_to_autocluster}
\alias{assign_eventreport_to_autocluster}
\title{Assign a single event report to an autodetected cluster on the database.}
\usage{
assign_eventreport_to_autocluster(autodetected_cluster_id, event_report_id)
}
\arguments{
\item{autodetected_cluster_id}{Id of the autodetected cluster (generally newly created).}

\item{event_report_id}{Id of the event report to assign.}
}
\description{
\code{assign_eventreport_to_autocluster} assigns an event report to an autodetected cluster on the database.
}
