% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{assign_initalsets_to_users}
\alias{assign_initalsets_to_users}
\title{Assigns initial article sets to user}
\usage{
# to assign training set and test set to user 3
assign_initialsets_to_users(con, 3)

# to asssing training set and test set to users 3,5, and 6
assing_initalsets_to_users(con, c(3, 5, 6))
}
\arguments{
\item{con}{The connection to the election violence database.}

\item{user_id}{Either a single user id or a vector of user ids.}
}
\description{
Assigns the training set and test sets to a user or set of users (articles which are already allocated to a user are not allocated again).
}
