% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{assign_reallocation_to_user}
\alias{assign_reallocation_to_user}
\title{Assign single events to user for reallocation to clusters.}
\usage{
assign_reallocation_to_user(event_report_id, user_id,
  allocation_date = as.character(Sys.Date()),
  last_updated = as.character(Sys.time()), completed = 0)
}
\arguments{
\item{event_report_id}{event_report_id of the single events.}

\item{user_id}{Id of the user the the cluster is to be assigned to.}

\item{allocation_date}{Date allocation made (usually today).}

\item{last_updated}{Date record last updated (usually today).}

\item{completed}{Is the cluster verification complete (usually 0 [i.e. not complete])}
}
\description{
\code{assign_reallocation_to_user} assigns an autodetected cluster to a user for verification.
}
