% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{assign_reallocations_to_user}
\alias{assign_reallocations_to_user}
\title{Assigns mutiple single events for reallocation to clusters to a single user.}
\usage{
assign_reallocations_to_user(event_report_ids, user_id)
}
\arguments{
\item{event_report_ids}{The autocluster_ids to allocate.}

\item{user_id}{The user id to allocate to.}
}
\description{
Assigns mutiple single events for reallocation to clusters to a single user.
}
