% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{assign_set}
\alias{assign_set}
\title{Assigns set of articles to a user}
\usage{
assign_set(con, user_id, set, allocation_type, allocated_by = "assign_set",
  status = "NEW")
}
\arguments{
\item{con}{The connection to the election violence database.}

\item{user_id}{The user id.}

\item{set}{A vector which contains the document ids of the articles to be allocated (generally the test set of articles and the training set of articles).}

\item{allocation_type}{The type of the allocation in the database (one of testing, training, coding, checking and ideal)}

\item{allocated_by}{The person (user_id) or the function which performed the allocation}
}
\description{
\code{assign_set} assigns a whole set of articles to a user (articles which are already allocated to a user are not allocated again).
This function is a convenience wrapper around \code{assign_article_to_user}.
}
\seealso{
\code{\link{assign_article_to_user}}
}
