% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{assign_testset_to_user}
\alias{assign_testset_to_user}
\title{Assign test set}
\usage{
assign_testset_to_user(con, user_id, testset = define_testset(),
  allocation_type = "testing", allocated_by = "assign_testset_to_user")
}
\arguments{
\item{con}{The connection to the election violence database.}

\item{user_id}{The user id.}
}
\description{
\code{assign_testset_to_user} assigns the test set of articles to a user specified by the user id. Articles which are already allocated to a user are not allocated again.
}
