% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{assign_trainingset_to_user}
\alias{assign_trainingset_to_user}
\title{Assigns the training set to user (articles which are already allocated to a user are not allocated again).}
\usage{
assign_trainingset_to_user(con, user_id, trainingset = define_trainingset(),
  allocation_type = "training", allocated_by = "assign_trainingset_to_user")
}
\arguments{
\item{con}{The connection to the election violence database.}

\item{user_id}{The user id.}
}
\description{
Assigns the training set to user (articles which are already allocated to a user are not allocated again).
}
