% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{assign_verified_clusters_to_user}
\alias{assign_verified_clusters_to_user}
\title{Assigns multiple verified clusters to a single user for combination.}
\usage{
assign_verified_clusters_to_user(verified_cluster_ids, user_id)
}
\arguments{
\item{verified_cluster_ids}{The verified_cluster_ids to allocate.}

\item{user_id}{The user id to allocate to.}
}
\description{
Assigns multiple verified clusters to a single user for combination.
}
