% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_analysis_patrick.R
\name{check_sngrams}
\alias{check_sngrams}
\title{Check whether a tokenized corpus contains specific n-grams. The function prints a list of the specific n-gram in all texts of the corpus}
\usage{
check_sngrams(toks, wseq, wseqc)
}
\arguments{
\item{toks}{Tokonized corpus}

\item{wseq}{Pre-specified word sequence as list on which n-grams should be created; may contain wildcard combinations using '*'}

\item{wseqc}{Pre-specified word sequence concatinated}
}
\description{
Check whether a tokenized corpus contains specific n-grams. The function prints a list of the specific n-gram in all texts of the corpus
}
