% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_analysis.R
\name{classifier_selection_description}
\alias{classifier_selection_description}
\title{Classifies new documents on a labeled training set (description).}
\usage{
classifier_selection_description(train, new_docs,
  text_field = "description", class_to_keep = 1,
  training_classify_var = "EV_article", prior = "uniform",
  classifier_type = "xgboost", stem = TRUE, return_logical = FALSE,
  ...)
}
\arguments{
\item{train}{a data frame with the training documents.}

\item{new_docs}{the documents to classify.}

\item{text_field}{the text field (must be the same in the training documents and the documents to classify).}

\item{class_to_keep}{the class (0 or 1) to keep.}

\item{training_classify_var}{the variable containing the labels in the training set.}

\item{prior}{for naive bayes classifier only.}

\item{classifier_type}{which classifier to use (xgboost or nb (naive Bayes))}

\item{stem}{for preprocessing}

\item{return_logical}{return the subset of documents or a logical vector indicating that subset.}

\item{...}{other arguments to be passed to \code{preprocess_corpus}.}
}
\description{
Classifies new documents on a labeled training set (description).
}
