% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_analysis.R
\name{classifier_selection_ocr}
\alias{classifier_selection_ocr}
\title{Classifies new documents on a labeled training set (description)}
\usage{
classifier_selection_ocr(train, new_docs, text_field = "ocr",
  class_to_keep = 1, training_classify_var = "EV_article",
  prior = "uniform", return_logical = FALSE)
}
\description{
Classifies new documents on a labeled training set (description)
}
