% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_scoring_grouping_method.R
\name{compare_summaries}
\alias{compare_summaries}
\title{Compares summaries of coding by user.}
\usage{
compare_summaries(coding_summary)
}
\arguments{
\item{coding_summary}{A coding summary (usually generated by the \code{coding_summary} function.)}
}
\description{
Compares summaries of coding by user.
}
