% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_output.R
\name{compile_report}
\alias{compile_report}
\title{Produces a readable summary of event reports from the election violence database.}
\usage{
compile_report(
  event_report_id,
  data_location = "database",
  output_format = "html_document",
  output_file = NULL,
  date_report_format = "\%A, \%d \%B \%Y",
  report_title = "Event Reports from Election Violence Database"
)
}
\arguments{
\item{event_report_id}{the event report id or ids to produce the report for.}

\item{data_location}{where data is retrived from (currently only database supported).}

\item{output_format}{produces reports as html_document (default), word_document, or pdf_document.}

\item{output_file}{the file name to save the report to (optional - if not supplied a temp file will be created).}

\item{date_report_format}{desired date format for output.}
}
\description{
Produces a readable summary of event reports from the election violence database.
}
