% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firststage_functions.R
\name{create_docsets}
\alias{create_docsets}
\title{Create training and test set document matrices}
\usage{
create_docsets(alldocs, size, seed)
}
\arguments{
\item{alldocs}{The document data.frame to be split}

\item{size}{Size of the training set (must be smaller than number of documents in the data.frame)}

\item{seed}{Seed for randomization to ensure replicability}
}
\description{
Create training and test set document matrices
}
