% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{create_new_autocluster}
\alias{create_new_autocluster}
\title{\code{create_new_autocluster} Creates a new autodetected cluster on the data.}
\usage{
create_new_autocluster()
}
\value{
the id on the database of the newly created autodetected cluster.
}
\description{
\code{create_new_autocluster} Creates a new autodetected cluster on the data.
}
