% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_analysis_patrick.R
\name{create_sets}
\alias{create_sets}
\title{Create training and test set from a corpus}
\usage{
create_sets(the_corpus, size, seed)
}
\arguments{
\item{the_corpus}{The text corpus to split}

\item{size}{Size of the training set (must be smaller than number of documents in the corpus)}

\item{seed}{Seed for randomization to ensure replicability}
}
\description{
Create training and test set from a corpus
}
