% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{date_to_boundary_year}
\alias{date_to_boundary_year}
\title{Returns the set of electoral boundaries that were in place given a year}
\usage{
date_to_boundary_year(x)
}
\arguments{
\item{x}{the date to check (yyyy-mm-dd format)}
}
\description{
Returns the set of electoral boundaries that were in place given a year
}
