% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_sets.R
\name{date_to_election}
\alias{date_to_election}
\title{Finds the election relating to any specific date.}
\usage{
date_to_election(the_date, election_start_col = "monthsearch_start",
  election_end_col = "monthsearch_end",
  election_data = durhamevp::election_dates,
  return_col = "election_name")
}
\arguments{
\item{the_date}{the date to classifiy}

\item{election_start_col}{the column in the election data frame which identifies when the election starts.
@param election_end_col the column in the election data frame which identifies when the election ends.
@param election_data the data frame containing the election data.
@param return_col the column in the data frame containing the name of the election (to be returned by the function).
@export}
}
\description{
Finds the election relating to any specific date.
}
