% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_sets.R
\name{define_testset}
\alias{define_testset}
\title{Returns the test set document ids.}
\usage{
define_testset(testset = 18:37)
}
\arguments{
\item{testset}{Vector of items which can replace the default test set (items 18 to 37).}
}
\description{
Returns the test set document ids.
}
