% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_sets.R
\name{define_trainingset}
\alias{define_trainingset}
\title{Returns the training set document ids.}
\usage{
define_trainingset(trainingset = 2:17)
}
\arguments{
\item{trainingset}{Vector of items which can replace the default training set (items 2 to 17).}
}
\description{
Returns the training set document ids.
}
