% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{documents_to_actual}
\alias{documents_to_actual}
\title{Restrict a vector to document ids actually existing in the database.}
\usage{
documents_to_actual(document_id)
}
\description{
Restrict a vector to document ids actually existing in the database.
}
