% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_analysis.R
\name{documents_to_latex}
\alias{documents_to_latex}
\title{Convert dataframe of documents to form suitable for asis printing in knitr or Rmarkdown documents.}
\usage{
documents_to_latex(out_docs, include_description = TRUE,
  include_ocr = TRUE, pandoc_output = FALSE)
}
\arguments{
\item{out_docs}{data frame containing the following fields: id, title, description (optional), ocr (optional)}

\item{include_description}{include description field in output}

\item{include_ocr}{include ocr field in output}

\item{pandoc_output}{should function produce pandoc output for Rmarkdown (if FALSE the output is LaTex output for knitr).}
}
\description{
Convert dataframe of documents to form suitable for asis printing in knitr or Rmarkdown documents.
}
