% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{download_to_superwide}
\alias{download_to_superwide}
\title{Transforms the downloaded coding to 'superwide' format}
\usage{
download_to_superwide(
  evp_coding_download,
  coding_mode_only = TRUE,
  coding_complete_only = TRUE
)
}
\arguments{
\item{evp_coding_download}{The result of a \code{get_coding()}.}

\item{coding_mode_only}{Whether to include only results from coding undertaken in by users in 'coding mode'.}

\item{coding_complete_only}{Whether to include only results where coding is marked as complete.}
}
\description{
Superwide format creates one row for every event_report. It then creates a separate variable for every table, variable, value combination and counts the number of instances associated with event report. For example the event report has one Conservative perpetrator then the superwide record variable will be actors_perpetrator_cons_tor_union and the superwide value for that variable will be 1.
Currently the locations are processed with a hack - choosing the first location of each type associated with an event report.
}
