% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ev_deaths}
\alias{ev_deaths}
\title{Election Violence Deaths data.}
\format{
A data frame with 96 rows and 26 variables, which are:
\describe{
\item{event_id}{The event_id the record relates to in the ev_events and event_reports dataframes.}
\item{election_name}{Identifies the election the record relates to.}
\item{case_id}{Death record within event: counter of the number of deaths within an event.}
\item{location}{String recording the location at which the death occurred.}
\item{victim_name}{Name of the victim if provided in the source material; otherwise NA.}
\item{victim_gender}{Gender of the victim if provided. Values: male, female, NA}
\item{victim_age}{Age in number of years if provided in the surce material; otherwise NA.}
\item{victim_age_text}{String providing the in-text description of the victim's age given; otherwise NA.}
\item{victim_profession}{String variable indicating the profession of the vitim if provided; otherwise NA.}
\item{vicim_election_role}{String indicating the role of the victim during the election if provided; otherwise NA.}
\item{victim_affiliation}{String recording the victim's party affiliation if provided; otherwise NA.}
\item{specific_circumstance}{String describing the specific circumstances of the death; otherwise NA.}
\item{grouped_circumstance}{String variable grouping circumstance of deaths into five categories. Values: Riot, Roughs, Candidate Meeting, Political Quarrel, and Other.}
\item{weapon}{String listing weapons used in the incident. If not provided in the source material NA is coded.}
\item{cause_death}{String recording the denoted cause of death if provided; otherwise NA.}
\item{election_point_death}{String indicating the timing of the death during the election proceedings. Values: during campaigning, during polling, after polling.}
\item{perpetrator}{String describing the perpetrator if provided in source material; otherwise NA.}
\item{perpetrator_affiliation}{String recording the perpetrator's party affiliation if provided; otherwise NA.}
\item{consequences}{String recording any electoral consequences of the death as noted n the source material; otherwise NA.}
\item{best_bna_link}{Link to the newspaper source inthe British Newspaper Archive with the most detailed account of the death. Links the other source material can be obtained by linking ev_deaths to event_reports and user_docs.}
}
}
\usage{
ev_deaths
}
\description{
The death data was created from the original coding. All events with at least one report involving a killing were individually checked to verify whether a death had actually occurred. If so, additional variables (listed below) have been extracted from the original coding sources. In total we found 96 individual deaths occurring during the 20 General Elections, 1832-1914. Data on deaths is added to the package in two ways: (1) some aggregate descriptives at the event level (see ev_events description) and (2) as stand-alone ev_deaths data frame, which is described below.
}
\keyword{datasets}
