% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{evdb_connect}
\alias{evdb_connect}
\title{Set connection to election violence database}
\usage{
evdb_connect(host = "coders.victorianelectionviolence.uk",
  user = "data_writer", dbname = "evp", port = 3306,
  password_method = "ask")
}
\arguments{
\item{host}{Location of database host.}

\item{user}{Username for accessing the database.}

\item{dbname}{Name of database.}

\item{port}{Value of port connection.}

\item{password_method}{Method of obtaining database password for specified user.}
}
\value{
The connection to the election violence database.
}
\description{
Set connection to election violence database
}
\examples{
# for default connection to the database
con <- evdb_connect()
# using alternative method of entering password information
con <- evdb_connect(password_method="config")
}
