% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_analysis.R
\name{evp_classifiers}
\alias{evp_classifiers}
\title{wrappers round classifiers}
\usage{
evp_classifiers(train_dfm, classifier_type, training_classify_var, prior)
}
\description{
wrappers round classifiers
}
