% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_scoring_utility.R
\name{extract_coding_matches}
\alias{extract_coding_matches}
\title{Creates tibble of all matches from the election violence database}
\usage{
extract_coding_matches()
}
\value{
A tibble of all the database entries which have been matched. The columns of the tibble are as follows \code{pair_no} unique id for the user_doc allocations (i.e. identifies the two user allocations at the document allocation level. \code{case_no} a unique id at the level one up from the match (i.e. identifying the two tag_ids for attributes, identifying the two event_report ids for tags, identifying the two document allocations for event reports). \code{model_var} is the id of the first record being compared, \code{user_var} is the id of the second record being compared, \code{match_type} is the database table the match is from (values: "user_doc", "ev_report", "tags", "attributes"), \code{score} is the matching score on which the match is based.
}
\description{
\code{extract_matches} is a wrapper function for applications of record matching.
}
