% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{find_const}
\alias{find_const}
\title{Finds the constituency (Vison of Britian G_NAME) of a latitude and longitude.}
\usage{
find_const(lat, lon, boundary_label,
  constituencies_shapes_list = durhamevp::const_shape_list,
  col_to_return = "G_NAME")
}
\arguments{
\item{lat}{the latitude to check.}

\item{lon}{the longitude to check.}

\item{boundary_label}{The name in the list of the Simple Feature dataframe to use.}

\item{constituencies_shapes_list}{The constituencies. A named list of Simple Features dataframes. The names identify which Simple Features Dataframe to use.}
}
\description{
Finds the constituency (Vison of Britian G_NAME) of a latitude and longitude.
}
