% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{find_county}
\alias{find_county}
\title{Finds county of a latitude and longitude.}
\usage{
find_county(lat, lon, county_shapes = durhamevp::county_shapes,
  col_to_return = "G_NAME")
}
\arguments{
\item{lat}{the latitude to check.}

\item{lon}{the longitude to check.}

\item{county_shapes}{A Simple Feature dataframe of the counties}
}
\description{
Finds the county (Vision of Britian G_NAME) of a location given by latitude and longitude.
}
