% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{get_allocation_connect_to_docs}
\alias{get_allocation_connect_to_docs}
\title{Get the allocation of articles to users (the unit of coding), with connected information from documents table and search tables}
\usage{
get_allocation_connect_to_docs(user_id = "all",
  allocation_type = "all", document_id = "all", user_doc_id = "all",
  include_ocr = FALSE)
}
\arguments{
\item{user_id}{The user_id to check in the database.}

\item{document_id}{The set of document_ids to include in the allocation}

\item{include_ocr}{Should the OCR of the document be included in the results.}
}
\value{
dataframe of the document allocations to the user.
}
\description{
Get the allocation of articles to users (the unit of coding), with connected information from documents table and search tables
}
