% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{get_archivesearches}
\alias{get_archivesearches}
\title{Returns the table of searches which have been passed to the crawler.}
\usage{
get_archivesearches(archive_search_id = "all")
}
\arguments{
\item{archive_search_id}{archive_search id or ids to filter by.}
}
\description{
Returns the table of searches which have been passed to the crawler.
}
