% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{get_archivesearchresults}
\alias{get_archivesearchresults}
\title{Returns the crawler search table}
\usage{
get_archivesearchresults(archivesearchresults_id = "all",
  archive_search_id = "all")
}
\arguments{
\item{archivesearch_id}{archive_search id or ids to filter by.}
}
\description{
Returns the crawler search table
}
