% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{get_archivesearchsummaryonly}
\alias{get_archivesearchsummaryonly}
\title{Returns the archivesearchessummaryonly table filtered by id}
\usage{
get_archivesearchsummaryonly(archivesearchsummaryonly_id = "all")
}
\arguments{
\item{event_report_id}{Event report id or ids to filter by.}
}
\description{
Returns the archivesearchessummaryonly table filtered by id
}
