% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{get_attribute}
\alias{get_attribute}
\title{Returns the tag table filtered by either attribute_id (primary key) or tag_id}
\usage{
get_attribute(attribute_id = "all", tag_id = "all")
}
\arguments{
\item{attribute_id}{attribute id or ids to filter by.}

\item{tag_id}{Foreign key from tag table}
}
\description{
Returns the tag table filtered by either attribute_id (primary key) or tag_id
}
