% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_autodetected_cluster_allocations}
\alias{get_autodetected_cluster_allocations}
\title{Get the autodetcted cluster allocations currently allocated to the user.}
\usage{
get_autodetected_cluster_allocations(user_id = "all",
  autodetected_cluster_id = "all")
}
\arguments{
\item{user_id}{The user_id to check in the database.}

\item{autodetected_cluster_id}{The set of autodetected cluster ids to check in the database}
}
\value{
dataframe of the autodetected cluster allocations to the user.
}
\description{
Get the autodetcted cluster allocations currently allocated to the user.
}
