% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_analysis.R
\name{get_candidates_fromarchivesearchresults}
\alias{get_candidates_fromarchivesearchresults}
\title{Gets candidate documents from archivesearchresults.
This is somewhat challenging because sometimes the url has 'download' in sometimes it doesn't
This function is really a hack to find either case.}
\usage{
get_candidates_fromarchivesearchresults(archivesearchresults)
}
\arguments{
\item{archivesearchresults}{The archive search results (including a url column)}
}
\value{
Candidate documents with urls matching the urls in archivesearchresults
}
\description{
Gets candidate documents from archivesearchresults.
This is somewhat challenging because sometimes the url has 'download' in sometimes it doesn't
This function is really a hack to find either case.
}
