% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_analysis.R
\name{get_classified_docs}
\alias{get_classified_docs}
\title{Function to get documents from database which are classified for text analysis.}
\usage{
get_classified_docs()
}
\description{
Function to get documents from database which are classified for text analysis.
}
