% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_combination_clusters}
\alias{get_combination_clusters}
\title{Finds combination clusters.}
\usage{
get_combination_clusters()
}
\description{
\code{get_combination_clusters} reports the verified clusters and the event reports currently in that cluster.
}
