% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_scoring_utility.R
\name{get_distance}
\alias{get_distance}
\title{Gets the Levenshtein distance between two separate rows after serializing them.}
\usage{
get_distance(model, user, type = "primitive")
}
\arguments{
\item{model}{the first row}

\item{user}{the second row}

\item{type}{the type of comparison to conduction (implying which fields are included in the comparison)}
}
\description{
Gets the Levenshtein distance between two separate rows after serializing them.
}
