% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{get_event_report}
\alias{get_event_report}
\title{Returns the event_report table filtered by event report id}
\usage{
get_event_report(event_report_id)
}
\arguments{
\item{event_report_id}{Event report id or ids to filter by.}
}
\description{
Returns the event_report table filtered by event report id
}
