% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_scoring_utility.R
\name{get_similarity_score}
\alias{get_similarity_score}
\title{Gets a similarity score between two strings}
\usage{
get_similarity_score(a, b)
}
\arguments{
\item{a}{string 1}

\item{b}{string 2}
}
\value{
Returns a similarity score between 0.0 (completely identical) and 1.0 (completely different)
}
\description{
Gets a similarity score between two strings
}
