% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_scoring_utility.R
\name{get_sorted}
\alias{get_sorted}
\title{Uses a scoring matrix to decide which model (model answer) component should be used to score
every user (user answer) component
throughout the function, 'lowest' scores are the ones that represent a near-perfect match}
\usage{
get_sorted(model_vars, user_vars, get_var_score = get_distance,
  blank_var_gen = NULL)
}
\arguments{
\item{model_vars}{dataframe containing the first set of codings to be matched}

\item{user_vars}{dataframe containing the second set of codings to be matched}

\item{get_var_score}{the function to do the comparison}
}
\description{
Uses a scoring matrix to decide which model (model answer) component should be used to score
every user (user answer) component
throughout the function, 'lowest' scores are the ones that represent a near-perfect match
}
