% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{get_tag}
\alias{get_tag}
\title{Returns the tag table filtered by either tag_id (primary key) or event report id}
\usage{
get_tag(tag_id = "all", event_report_id = "all")
}
\arguments{
\item{tag_id}{Primary key from tag table}

\item{event_report_id}{Event report id or ids to filter by.}
}
\description{
Returns the tag table filtered by either tag_id (primary key) or event report id
}
