% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{get_user}
\alias{get_user}
\title{Gets user details.}
\usage{
get_user(con, user_id)
}
\arguments{
\item{con}{The connection to the election violence database.}

\item{user_id}{The user id.}
}
\description{
Gets user details.
}
