% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{get_user_mode}
\alias{get_user_mode}
\title{Returns table of current user modes}
\usage{
get_user_mode(user_id = "all")
}
\arguments{
\item{user_id}{Id of the user whose mode to check or a vector of user ids. The default ("all") selects all users.}
}
\description{
Returns the table of current user modes.
}
