% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{get_useractivitylogentry}
\alias{get_useractivitylogentry}
\title{Get user logged activity}
\usage{
get_useractivitylogentry(useractivitylogentry_id = "all",
  action = "all", user_id = "all", allocation_id = "all")
}
\arguments{
\item{useractivitylogentry_id}{primary key of user activity table of interest.}

\item{action}{action types of interest.}

\item{user_id}{Id of the user whose activity is of interest.}

\item{allocation_id}{allocation_ids of interest.}
}
\description{
Gets the log of user activity on the database relating to a user or an allocation or an action type.
}
