% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_verified_clusters}
\alias{get_verified_clusters}
\title{Finds verified clusters (and their existing size).}
\usage{
get_verified_clusters()
}
\description{
\code{get_verified_clusters} reports the verified clusters and the event reports currently in that cluster.
}
