% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_analysis.R
\name{has_word}
\alias{has_word}
\title{Function to tally cases with and without a particular words}
\usage{
has_word(
  the_dataframe,
  contains_words,
  text_col = "ocr",
  results_col = "EV_article"
)
}
\arguments{
\item{the_dataframe}{the dataframe with a text column and a results column}

\item{contains_words}{the word (or other regex) to search for}

\item{text_col}{the column containing the text}

\item{results_col}{the column containing the classification}
}
\description{
Function to tally cases with and without a particular words
}
