% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{is_evp_coding_download}
\alias{is_evp_coding_download}
\title{Check if object is an evp coding download.}
\usage{
is_evp_coding_download(coding_download)
}
\arguments{
\item{coding_download}{object to check}
}
\value{
logical
}
\description{
Check if object is an evp coding download.
}
