% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{is_location_tags}
\alias{is_location_tags}
\title{Checks if an object is a location_tags table}
\usage{
is_location_tags(location_tags)
}
\arguments{
\item{location_tags}{object to check}
}
\description{
Checks if an object is a location_tags table
}
