% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{keyring_evp_available}
\alias{keyring_evp_available}
\title{Checks if an evp password is available.}
\usage{
keyring_evp_available(username = "not implemented")
}
\description{
Checks if an evp password is available on this computer.
}
