% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{manage_dbcons}
\alias{manage_dbcons}
\title{Returns a connection to the election violence database}
\usage{
manage_dbcons(host = "coders.victorianelectionviolence.uk",
  username = "data_writer", dbname = "evp", port = 3306)
}
\description{
This function is primarily for use within other functions.
This is now a wrapper function around pool::dbPool to handle connections to the database.
Database password is obtained from the keyring if there is a password for the database in the keyring. Otherwise the user is prompted for the database password.
}
