% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geography.R
\name{map_points_on_background}
\alias{map_points_on_background}
\title{plots sf points against background of a polygons (the_shapes)}
\usage{
map_points_on_background(z, the_shapes, title = "my map",
  labels = TRUE)
}
\arguments{
\item{z}{the points (sf format).}

\item{the_shapes}{the background polygons.}

\item{title}{the title of the map.}

\item{labels}{should the points be labelled with their G_NAME.}
}
\description{
plots sf points against background of a polygons (the_shapes)
}
