% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/define_sets.R
\name{month_contains_election}
\alias{month_contains_election}
\title{Finds the election relating to a month (where an election took place on any day that month).}
\usage{
month_contains_election(the_date,
  election_start_col = "monthsearch_start",
  election_end_col = "monthsearch_end",
  election_data = durhamevp::election_dates,
  return_col = "election_name")
}
\arguments{
\item{the_date}{the month to classifiy as yyyy-mm-dd where dd must be 01.}

\item{election_start_col}{the column in the election data frame which identifies when the election starts.}

\item{election_end_col}{the column in the election data frame which identifies when the election ends.}

\item{election_data}{the data frame containing the election data.}

\item{return_col}{the column in the data frame containing the name of the election (to be returned by the function).}
}
\description{
Finds the election relating to a month (where an election took place on any day that month).
}
