% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firststage_functions.R
\name{nb_keywords}
\alias{nb_keywords}
\title{Train a naive bayes classifier and extract keywords as data.frame object ordered by most predictive word of class on top (wrapper function around Quanteda commands). The returned data.frame has three columns: word, 0(=posterior prob of word indicating that document in classvar==0), 1(=posterior prob of word predicting word indicating that document in classvar==1), and id.}
\usage{
nb_keywords(training, classvar, distribution = "Bernoulli")
}
\arguments{
\item{training}{Training data feature matrix}

\item{classvar}{Classification variable name, entered as string}
}
\description{
Train a naive bayes classifier and extract keywords as data.frame object ordered by most predictive word of class on top (wrapper function around Quanteda commands). The returned data.frame has three columns: word, 0(=posterior prob of word indicating that document in classvar==0), 1(=posterior prob of word predicting word indicating that document in classvar==1), and id.
}
