% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_analysis_patrick.R
\name{nb_test}
\alias{nb_test}
\title{Train a naive bayes classifier on a training dfm and asses its performance on a test dfm reporting test statistics (wrapper function aroudn Quanteda commands)}
\usage{
nb_test(training, testing, classvar, distribution = "Bernoulli")
}
\arguments{
\item{training}{Training data frequency matrix}

\item{testing}{Testing data frequency matrix}

\item{classvar}{Classification variable name, entered as string}
}
\description{
Train a naive bayes classifier on a training dfm and asses its performance on a test dfm reporting test statistics (wrapper function aroudn Quanteda commands)
}
