% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_analysis.R
\name{preprocess_corpus}
\alias{preprocess_corpus}
\title{Preprocess a text corpus and return a document feature matrix (wrapper round quanteda functions).}
\usage{
preprocess_corpus(the_corpus, min_termfreq = 2, min_docfreq = 2,
  remove_punct = TRUE, remove_numbers = TRUE, remove_hyphens = TRUE,
  dfm_tfidf = FALSE)
}
\arguments{
\item{the_corpus}{The text corpus to split.}
}
\description{
Preprocess a text corpus and return a document feature matrix (wrapper round quanteda functions).
}
