% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_analysis_patrick.R
\name{preprocess_ngrams}
\alias{preprocess_ngrams}
\title{Preprocess a text corpus including the creation of n-grams and return a document feature matrix (wrapper round quanteda functions).}
\usage{
preprocess_ngrams(the_corpus, n, min_termfreq = 2, min_docfreq = 2,
  max_termfreq = NULL, max_docfreq = NULL, remove_punct = TRUE,
  remove_numbers = TRUE, remove_hyphens = TRUE,
  termfreq_type = "count", docfreq_type = "count", dfm_tfidf = FALSE)
}
\arguments{
\item{the_corpus}{The text corpus to be pre-processed.}

\item{n}{Upper-bound of n-grams to be included. E.g., entering 2 would mean that uni-grams and bi-grams are included}
}
\description{
Preprocess a text corpus including the creation of n-grams and return a document feature matrix (wrapper round quanteda functions).
}
