% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{process_locations}
\alias{process_locations}
\title{Creates a wide version of the locations associated with event reports. Currently based on a hack.}
\usage{
process_locations(location_tags)
}
\arguments{
\item{location_tags}{a table of tags (filtered: table=="location) containing only locations.}
}
\description{
Creates a wide version of the locations associated with event reports. Currently based on a hack.
}
