% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article_assignment.R
\name{reallocate_from_df}
\alias{reallocate_from_df}
\title{Assign users to documents from a dataframe}
\usage{
reallocate_from_df(reassign_dat, allocation_type, allocated_by)
}
\arguments{
\item{reassign_dat}{A dataframe (or tibble) with a user_id column and a document_id column}

\item{allocation_type}{The allocation type to write to the database}

\item{allocated_by}{The allocated by to write to the database}
}
\description{
Assign users to documents from a dataframe where dataframe has a user_id column and and a document_id column
}
