% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/article_assignment.R
\name{reallocate_randomly}
\alias{reallocate_randomly}
\title{Randomly reassign a set of articles amongst a set of users for coding}
\usage{
reallocate_randomly(user_ids, user_doc_ids,
  allocated_by = "random_reassignment", restrict_to_actual = TRUE,
  make_assignments = TRUE, force = FALSE)
}
\arguments{
\item{user_ids}{The users to allocate the articles amongst (vector with single or multiple user_ids).}

\item{user_doc_ids}{The set of documents which are to be allocated (vector with single or multiple of document_ids).}

\item{allocated_by}{The value to write to the allocated_by field in the database document_allocations table.}

\item{restrict_to_actual}{Should the restriction to actual users and documents be enforced. Should only be set to FALSE for debugging purposes.}

\item{make_assignments}{Actually make changes to the database (TRUE) or only create a proposed set of changes as a dataframe (FALSE)}

\item{force}{Make changes even if usual conditions are violated (status=="NEW", allocation_type=="coding")}
}
\description{
\code{reallocate_randomly} reassigns a set of already existing articles amongst a set of users for them to code in the election violence database.
By default allocation_ids and user ids are checked against the database and codes which do not correspond to existing assignments and/or users are ignored.
}
