% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{remove_evp_db_pool}
\alias{remove_evp_db_pool}
\title{Remove database pool}
\usage{
remove_evp_db_pool()
}
\description{
Removes hidden election violence database pool from R Global Environment by assinging it a NULL value. Use this function if the db pool is returning errors and a new pool will be created next time the database is queried.
@export
}
