% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firststage_functions.R
\name{run_firststage_fcm}
\alias{run_firststage_fcm}
\title{First stage function of the article selection process taking in a set of documents and returning a set of potential keywords}
\usage{
run_firststage_fcm(docs, docidvar = "fakeid", classvar = "classified",
  typevar = "EV_article", textvar = "description", stem = TRUE,
  min_termfreq = 20, min_docfreq = 20, max_termfreq = NULL,
  max_docfreq = NULL, remove_punct = TRUE, remove_numbers = TRUE,
  remove_hyphens = TRUE, termfreq_type = "count",
  docfreq_type = "count", dfm_tfidf = FALSE, initialkw = c("elect",
  "riot", "disturb", "incid"), cpoint2 = 0.9)
}
\arguments{
\item{docs}{Data frame of documents containing classified cases (R-set) and unclassified cases (S-set)}

\item{docidvar}{Unique document id variable; default = "fakeid"}

\item{classvar}{Indicator identifying classified documents; default = "classified"}

\item{typevar}{Indicator identifying election violence articles; default = "EV_article"}

\item{textvar}{Indicator identifying text field to classify on; default = "description"}

\item{stem}{default FALSE}

\item{min_termfreq}{default 20}

\item{min_docfreq}{default 20}

\item{max_termfreq}{default NULL}

\item{max_docfreq}{default NULL}

\item{remove_punct}{default TRUE}

\item{remove_numbers}{default TRUE}

\item{remove_hyphens}{default TRUE}

\item{termfreq_type}{default "count"}

\item{docfreq_type}{default "count"}

\item{dfm_tfidf}{default FALSE}

\item{initialkw}{Initial keywords used to retrive classified documents; default = c("elect", "riot", "disturb", "incid")}

\item{cpoint2}{Cutpoint on predicability of keyword in step 2; default = 0.9}
}
\description{
First stage function of the article selection process taking in a set of documents and returning a set of potential keywords
}
