% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_analysis.R
\name{searches_to_dfm}
\alias{searches_to_dfm}
\title{Convert a set of search results to a dfm.}
\usage{
searches_to_dfm(archivesearches)
}
\arguments{
\item{archivesearches}{the results of archivesearches}
}
\description{
Convert a set of search results to a dfm.
}
