% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_scoring_utility.R
\name{serialize_row}
\alias{serialize_row}
\title{Turns a database row into a string}
\usage{
serialize_row(df_row, serialization_type = "primitive")
}
\arguments{
\item{df_row}{the database row}

\item{serialization_type}{defines which columns to serialize ("primitive", "testing", "coding")}
}
\description{
Turns a database row into a string
}
