% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_analysis.R
\name{status_to_text}
\alias{status_to_text}
\title{Convert candidate document status to text}
\usage{
status_to_text(status)
}
\arguments{
\item{status}{status as recorded in ev database}
}
\description{
Convert candidate document status to text
}
