% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{user_allocated_fromset}
\alias{user_allocated_fromset}
\title{Returns the elements of a set of documents allocated (or not allocated) to a user.}
\usage{
user_allocated_fromset(con, user_id, set, not_allocated = FALSE)
}
\arguments{
\item{con}{The database connection to the election violence database.}

\item{user_id}{The user id to check in the database.}

\item{set}{The set of documents to check.}

\item{not_allocated}{If FALSE (default) returns the items allocated to the user, if TRUE returns items NOT allocated to the user.}
}
\value{
A vector numbers representing the document ids from the set which are allocated to the user (or which are not allocated to the user if not_allocated=TRUE).
}
\description{
Returns the elements of a set of documents allocated (or not allocated) to a user.
}
\examples{
# returns the trainingset documents allocated to user 1
user_allocated_fromset(con, 1, set=define_trainingset())

# returns the trainingset documents not allocated to user 1
user_allocated_fromset(con, 1, set=define_trainingset(), not_allocated=TRUE)


}
