% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{user_docs_partisanmerger}
\alias{user_docs_partisanmerger}
\title{Mitchell Newspaper Partisan Affiliation for Each user_doc.
The partisan information comes from the Mitchell Advertiser handbooks for 1847, 1856 (linked to the 1857 election), 1868, 1885, and 1910. The partisan leaning coding is described by the newspapers themselves in the Mitchell and was translated by Dr Gary Hutchison into two variables, partisan and leaning. Sources from our dataset were matched to the Mitchell data based on (1) publication title, (2) publication location, and (3) publication weekday/frequency. This ensures that newspapers with the same name are matched correctly and allows weekday and weekend editions to have different partisan leanings. This process was done by hand by a trained and supervised RA. The quality of the match is recorded in match_quality, where 3 means perfect match on all three criteria, 2 means there is good match on two criteria, and 1 means there is a match, but only on one criteria (generally title). Where match_quality==NA, no match could be found. The Mitchell data is added to the document_level data and missing years have been inter- and extrapolated using forward and backward panel completion. There are purely interpolated variables (only partisan and leaning_simple), where “lof” and “bw” indicates whether the last observation is brought forward or backward, and fully inter- and extrapolated variables. Those partisan and leaning_simple variables are named accordingly in the data.}
\format{
A data frame with 11667 rows and 14 variables, which are:
\describe{
\item{user_doc_id}{The user_doc_id the record relates to (links to user_doc table in the database).}
\item{mitchell_title}{The title of the publication as recorded in the Mitchell handbook.}
\item{mitchell_partisan} {The partisan leaning as recorded in the Mitchell handbook.}
\item{leaning} {A simplified version of the Mitchell partisan variable. Values: liberal, lib-union, lib-lab, lib-con, radical, con-lib, conservative, religious-conservative, repeal, netural, UNKNOWN, NA}
\item{partisan} {Records whether a newspaper has a partisan leaning or not. The variable can take there values “Y”=yes, “N”=no, “UNKNOWN”= unknown, i.e., the information given does not allow us to make a judgement if and whether a paper has a specific political leaning. NAs mean that no partisan information was recorded (if a paper could be matched) or no paper in the Mitchell list could be matched to the newspaper source.}
\item{partisan_interpol_f} {Interpolated partisan variable using last observation forwards. No extrapolation. Values: Y (yes), N (no), UNKNOWN, NA}
\item{partisan_interpol_b} {Interpolated partisan variable using last observation backwards. No extrapolation.. Values: Y (yes), N (no), UNKNOWN, NA}
\item{partisan_full_f} {Interpolated and extrapolated partisan variable using last observation forwards if possible, and last observation backwards if necessary.. Values: Y (yes), N (no), UNKNOWN, NA}
\item{partisan_full_b}{Interpolated and extrapolated partisan variable using last observation backwards if possible, and backwards if necessary.. Values: Y (yes), N (no), UNKNOWN, NA}
\item{leaning_simple} {Records the leaning of the paper. The variable can take three values: “conservative-right”, “liberal-left”, or “neutral”. “neutral” is only possible is partisan==N. NA means a leaning could not be determined if partisan==Y or partisan==NA if there was a match or simply that there was not Mitchell match.}
\item{leaning_simple_interpol_f}{Interpolated leaning_simple variable using last observation forwards. No extrapolation.}
\item{leaning_simple_interpol_b}{Interpolated leaning_simple variable using last observation backwards. No extrapolation.}
\item{leaning_simple_full_f}{Interpolated and extrapolated leaning_simple variable using last observation forwards if possible, and last observation backwards if necessary}
\item{leaning_simple_full_b}{Interpolated and extrapolated leaning_simple variable using last observation backwards if possible, and last observation forwards if necessary}
}
}
\usage{
user_docs_partisanmerger
}
\description{
Mitchell Newspaper Partisan Affiliation for Each user_doc.
The partisan information comes from the Mitchell Advertiser handbooks for 1847, 1856 (linked to the 1857 election), 1868, 1885, and 1910. The partisan leaning coding is described by the newspapers themselves in the Mitchell and was translated by Dr Gary Hutchison into two variables, partisan and leaning. Sources from our dataset were matched to the Mitchell data based on (1) publication title, (2) publication location, and (3) publication weekday/frequency. This ensures that newspapers with the same name are matched correctly and allows weekday and weekend editions to have different partisan leanings. This process was done by hand by a trained and supervised RA. The quality of the match is recorded in match_quality, where 3 means perfect match on all three criteria, 2 means there is good match on two criteria, and 1 means there is a match, but only on one criteria (generally title). Where match_quality==NA, no match could be found. The Mitchell data is added to the document_level data and missing years have been inter- and extrapolated using forward and backward panel completion. There are purely interpolated variables (only partisan and leaning_simple), where “lof” and “bw” indicates whether the last observation is brought forward or backward, and fully inter- and extrapolated variables. Those partisan and leaning_simple variables are named accordingly in the data.
}
\keyword{datasets}
