% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{users_to_actual}
\alias{users_to_actual}
\title{Restrict a vector to user ids actually existing in the the database.}
\usage{
users_to_actual(user_id)
}
\description{
Restrict a vector to user ids actually existing in the the database.
}
