% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{assign_article_to_user}
\alias{assign_article_to_user}
\title{Assign article to a user.}
\usage{
assign_article_to_user(con, document_id, user_id, allocation_date,
  allocation_type, allocated_by, status, coding_complete = 0)
}
\arguments{
\item{con}{The database connection to the election violence database.}

\item{document_id}{Id of the document to be assigned.}

\item{user_id}{Id of the user the document is to be assigned to.}

\item{allocation_date}{Date allocation made (usually today).}

\item{allocation_type}{Type of allocation (training, testing, coding, checking, ideal).}

\item{status}{Status of document coding (generally 'New' for newly assigned documents).}
}
\description{
\code{assign_article_to_user} assigns a specific article to a specific user.
}
