% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{assign_autocluster_to_user}
\alias{assign_autocluster_to_user}
\title{Assign autodetected cluster to a user.}
\usage{
assign_autocluster_to_user(autodetected_cluster_id, user_id,
  clusterattempt_id, allocation_date = as.character(Sys.Date()),
  last_updated = as.character(Sys.time()), completed = 0,
  needs_additional_checks = 0)
}
\arguments{
\item{autodetected_cluster_id}{Id of the autodetected cluster (generally newly created).}

\item{user_id}{Id of the user the the cluster is to be assigned to.}

\item{clusterattempt_id}{Id of the attempt at clustering (allows multiple clusterings)}

\item{allocation_date}{Date allocation made (usually today).}

\item{last_updated}{Date record last updated (usually today).}

\item{completed}{Is the cluster verification complete (usually 0 [i.e. not complete])}
}
\description{
\code{assign_autocluster_to_user} assigns an autodetected cluster to a user for verification.
}
