% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{assign_coding_to_environment}
\alias{assign_coding_to_environment}
\title{Assign the results of a get_coding download to the global environment}
\usage{
assign_coding_to_environment(evp_coding_download)
}
\arguments{
\item{evp_coding_download}{The result from executing the get_coding() command.}
}
\description{
Assign the results of a get_coding download to the global environment
}
