% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_analysis.R
\name{classifier_selection_keywords}
\alias{classifier_selection_keywords}
\title{\code{classifier_selection_keyword} uses a classifier to select document from keywords.
In 'select' mode the keywords are constructed from the archivesearchresults. In 'eval' mode the keywords are taken from the eval_options and a binary dfm is constructed from the document text.}
\usage{
classifier_selection_keywords(train, archivesearchresults, class_to_keep = 1,
  training_classify_var = "EV_article", prior = "docfreq",
  text_field = "ocr", classifier_type = "xgboost", mode = "select",
  eval_options = list(keywords = c("candidate", "poll", "election", "stone",
  "riot", "mob", "husting", "disturbance", "rough", "incident"), text_field =
  "ocr", eval_classify_var = "EV_article"))
}
\arguments{
\item{train}{the training set of documents}
}
\description{
\code{classifier_selection_keyword} uses a classifier to select document from keywords.
In 'select' mode the keywords are constructed from the archivesearchresults. In 'eval' mode the keywords are taken from the eval_options and a binary dfm is constructed from the document text.
}
