% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_scoring_grouping_method.R
\name{coding_summary}
\alias{coding_summary}
\title{Summary of coding of document allocation.}
\usage{
coding_summary(user_doc_id, allocation_type = "all",
  restrict_double_coded = TRUE)
}
\arguments{
\item{user_doc_id}{The user_doc_id(s) to summarize.}

\item{allocation_type}{The allocation type ("training", "testing","coding", "all").}

\item{restrict_double_coded}{Only summarize articles which have been coded more than once (in FALSE summarize all articles).}
}
\value{
Returns tibble with number of items of predefined types coded in the document by the user.
}
\description{
Returns tibble with number of items of each variable value combination types coded in the document by the user.
}
