% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{election_dates}
\alias{election_dates}
\title{Election Dates and Basic Information.}
\format{
A data frame with 20 rows and 11 variables:
\describe{
\item{election_name}{The standardized name of the election,used to link to other data sets in the election violence project.}
\item{keywords_13_complete}{Indicates whether the projects the BNA search using the standard 13 keywords to detect election violnece has been completed.}
}
}
\source{
Gary Hutchison collected the information from various sources
}
\usage{
election_dates
}
\description{
A dataset containing the dates and other basic information about British General Elections between 1832 and 1910.
}
\keyword{datasets}
