% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{get_candidate_documents}
\alias{get_candidate_documents}
\title{Returns the document table filtered by document id}
\usage{
get_candidate_documents(con, cand_document_id)
}
\arguments{
\item{con}{The connection to the election violence database.}

\item{cand_document_id}{Document id or ids to filter by.}
}
\description{
Returns the document table filtered by document id
}
