% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{get_candidate_documents}
\alias{get_candidate_documents}
\title{Returns the candidate document table filtered by candidate document id}
\usage{
get_candidate_documents(cand_document_id, url = "all",
  include_ocr = TRUE)
}
\arguments{
\item{cand_document_id}{Candidate document id or ids to filter by.}

\item{url}{document url or urls to filter by.}

\item{include_ocr}{should the ocr be downloaded.}
}
\description{
Note that excluding the ocr from the download will significantly reduce the time for the query to execute.
}
