% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_analysis.R
\name{get_candidates_fromarchivesearchresults}
\alias{get_candidates_fromarchivesearchresults}
\title{Gets candidate documents from archivesearchresults.
This is somewhat challenging because sometimes the url has 'download' in sometimes it doesn't
This function includes a hack to find either case.
By default it excludes articles from publications in Ireland and Scotland, and documents already classified
as 3 (verbatim repeat), 4 (Ireland), 5 (Scotland), 6 (Abroad)}
\usage{
get_candidates_fromarchivesearchresults(archivesearchresults,
  include_ocr = FALSE, restrict_EW = TRUE, restrict_classified = TRUE)
}
\arguments{
\item{archivesearchresults}{The archive search results (including a url column)}

\item{restrict_EW}{remove results published in Republic of Ireland and Scotland}

\item{restrict_classified}{remove results already classified as 3, 4, 5 or 6}

\item{include}{the ocr in the download (will slow down query)}
}
\value{
Candidate documents with urls matching the urls in archivesearchresults
}
\description{
Gets candidate documents from archivesearchresults.
This is somewhat challenging because sometimes the url has 'download' in sometimes it doesn't
This function includes a hack to find either case.
By default it excludes articles from publications in Ireland and Scotland, and documents already classified
as 3 (verbatim repeat), 4 (Ireland), 5 (Scotland), 6 (Abroad)
}
