% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_analysis.R
\name{get_classified_docs}
\alias{get_classified_docs}
\title{Function to get documents from database which are classified for text analysis.}
\usage{
get_classified_docs()
}
\value{
Returns a dataframe containing three types of article: first, general articles (unrelated to elections), second, some election but not violent articles and third, some election violence articles. The dataframe can be split into separate corpuses (if desired). The different types of article are distinguished using dummy variables: \code{EV_article} is \code{1} for election violence articles and \code{0} for all other articles. \code{election_article} is \code{1} for election articles (including election violence articles) and \code{0} for all other articles. (there is an unncessary \code{just_election} indicator for convenience which is \code{1} for election but not violence articles and \code{0} for election violence and general articles). The full ocr is in the field \code{ocr} and the short two line description is in the field \code{description}. The unique identifier column \code{fakeid} does not correspond to any id in the database because the data is aggregated from two different tables in the database (documents and candidate_documents).
}
\description{
\code{get_classified_docs} retrives the currently classified document set from the database. The documents have been classifed to enable machine learning.
}
