% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{get_clustering}
\alias{get_clustering}
\title{Download clustering process - all stages.}
\usage{
get_clustering()
}
\description{
\code{get_combined_clusters} reports the clusters after the combination stage and the event reports currently in that cluster.
}
