% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{get_coding}
\alias{get_coding}
\title{Download all the coding on the database}
\usage{
get_coding(include_ocr = FALSE)
}
\arguments{
\item{include_ocr}{Should results include the full ocr of the documents (will slow the download).}
}
\description{
Download all the coding on the database
}
