% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database_interactivity.R
\name{manage_dbcons}
\alias{manage_dbcons}
\title{Returns a connection to the election violence database}
\usage{
manage_dbcons()
}
\description{
This function is primarily for use within other functions.
Returns an active connection to the the election violence database, choosing the first open valid connection to the database. If there is no open valid connection to the database then a connection will be made. Running the function will also close all invalid (e.g. expired) database connections.
Database password is obtained from the keyring if there is a password for the database in the keyring. Otherwise the user is prompted for the database password.
}
