% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/firststage_functions.R
\name{preprocess_corpus}
\alias{preprocess_corpus}
\title{Preprocess a text corpus and return a document feature matrix (wrapper round quanteda functions).}
\usage{
preprocess_corpus(
  the_corpus,
  stem = TRUE,
  min_termfreq = 20,
  min_docfreq = 20,
  max_termfreq = NULL,
  max_docfreq = NULL,
  remove_punct = TRUE,
  remove_numbers = TRUE,
  remove_hyphens = TRUE,
  termfreq_type = "count",
  docfreq_type = "count",
  dfm_tfidf = FALSE
)
}
\arguments{
\item{the_corpus}{The text corpus to be pre-processed.}

\item{stem}{default TRUE}

\item{min_termfreq}{default 20}

\item{min_docfreq}{default 20}

\item{max_termfreq}{default NULL}

\item{max_docfreq}{default NULL}

\item{remove_punct}{default TRUE}

\item{remove_numbers}{default TRUE}

\item{remove_hyphens}{default TRUE}

\item{termfreq_type}{default "count"}

\item{docfreq_type}{default "count"}

\item{dfm_tfidf}{default FALSE}
}
\description{
Preprocess a text corpus and return a document feature matrix (wrapper round quanteda functions).
}
