% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_analysis.R
\name{split_corpus}
\alias{split_corpus}
\title{Divide a text corpus into training and testing sets based on number of training items}
\usage{
split_corpus(the_corpus, n_train)
}
\description{
Divide a text corpus into training and testing sets based on number of training items
}
