% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_analysis.R
\name{split_dfm}
\alias{split_dfm}
\title{Divide a document feature matrix into training and testing sets based on number of training items}
\usage{
split_dfm(dfm, n_train)
}
\arguments{
\item{dfm}{The document feature matrix to split.}

\item{n_train}{The number of documents in the training set.}
}
\description{
Divide a document feature matrix into training and testing sets based on number of training items
}
